requirements.txt
Flask==0.12
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
hdfs==2.0.16
Jinja2==2.9.6
Pillow==4.1.1
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.7
Werkzeug==0.12.1
WTForms==2.1
